dashboard/requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.3
packaging==24.1
pyserial==3.5
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
requests==2.32.3
scipy==1.10.1
simple-websocket==1.0.0
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
